Feature: Support BackChannel Authentication
This adds support for using backchannel authentication.
For example, your users authenticate using a public URL, but you are
adjacent to the auth server which you can reach via a K8s cluster.local
address. TLS is still required, but we skip certificate validation since
the cert cannot be signed for the back channel URL.

This also swaps out the gin-contrib session middleware for go-session.
Serializing the entire *oauth2.Token was with gob was surpassing the
MaxLength - we encode it to string now. This size limitation also
followed when using memcache or other storage options. go-session
doesn't have this size limit for other storage options, but still relies
on gorilla/securecookie for cookie storage.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>